# manifest: refactor Annotator with generics and add range annotations